Better websocket handling
Websocket worked fine for local development but two issues arise
as soon as the server and the client were on different machine.

1. The hostname is fixed to localhost, which won't work if the file
   are served from say 192.168.1.25

2. The port is fixed to 8081 which might cause issue if the server
   is asked to run on that port and will cause issue if the server
   machine didn't open the port 8081.

Both fix require editing the same line in the view and I stumbled
upon both issue at once so I fix both in the same commit. Current
solution for the port issue is to use the HTTP port + 1. Things stay
the same for the default case, solve the issue of starting express
on port 8081 and fix my issue as home server have port 5000 to 5999
open.